requirements.txt
pywin32>=306
pystray>=0.19.5
Pillow>=10.0.0
pynput>=1.7.6
